requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.2
django-cors-headers==4.3.1
django-environ==0.10.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.2
idna==3.6
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mysqlclient==2.2.3
oauthlib==3.2.2
pdfkit==1.0.0
pillow==10.2.0
pkgconfig==1.5.5
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
schema==0.7.5
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.2.0
xlwt==1.3.0
